requirements.txt
# Python 3.11
torch
datasets
transformers
sentencepiece
protobuf
python-dotenv
accelerate
fugashi
ipadic
bitsandbytes
unidic_lite
python2latex
colorspacious
matplotlib
prodigy-iaa
sqlalchemy
wandb
plotly
statsmodels
unsloth
